动态路由，404页面

+++ src/router/index.js
@@ -3,8 +3,8 @@ import store from "@/store/index"
 import base from "@/views/base"
 import login from "@/views/login/login"
 import workplace from "@/views/dashboard/workplace"
+import { injectionRouter } from "@/utils/commen"
 import error404 from "@/views/error/404"
-
 
 const routes = [
   {
@@ -19,9 +19,8 @@ const routes = [
         path: "/",
         name: "base",
         component: workplace,
-      },
-      {
-        path: '/404',
+      }, {
+        path: "/404",
         name: "404",
         component: error404
       }
@@ -31,12 +30,7 @@ const routes = [
     path: "/login",
     name: "login",
     component: login
-  },
-  {
-    path: '/:pathMatch(.*)*',
-    redirect: "/404",
-    hidden: true
-  },
+  }
 ]
 
 
@@ -46,18 +40,24 @@ const router = createRouter({
 })
 
 router.beforeEach((to, from, next) => {
+  const menuList = store.state.menu.menuList
+  injectionRouter(menuList, router)
+  debugger
+  if (!router.hasRoute(to.name) && to.name !=='404') {
+    next({
+      path: "/404"
+    })
+  }
   if (to.matched.some(r => r.meta.checkLogin)) {
     if (store.state.authorization) {
       next()
     } else {
       localStorage.clear()
       next({
-        path: "/login",
-        query: {redirect: to.fullPath}
+        path: "/login"
       })
     }
   } else {
-    console.log(next())
     next()
   }
 })
